requirements.txt
flake8==7.0.0
black==24.4.0
django==5.0.4
djangorestframework==3.15.1
django-extensions==3.2.3
python-decouple==3.8
psycopg2-binary==2.9.9
django-phonenumber-field[phonenumbers]==7.3.0
djangorestframework-simplejwt==5.3.1
ipdb==0.13.13
debugpy==1.8.1
redis[hiredis]==5.0.4
Faker==24.11.0
